+++ server/faker-gl.cpp
@@ -1,6 +1,6 @@
 // Copyright (C)2004 Landmark Graphics Corporation
 // Copyright (C)2005, 2006 Sun Microsystems, Inc.
-// Copyright (C)2009, 2011-2012, 2015, 2018-2020 D. R. Commander
+// Copyright (C)2009, 2011-2012, 2015, 2018-2021 D. R. Commander
 //
 // This library is free software and may be redistributed and/or modified under
 // the terms of the wxWindows Library License, Version 3.1 or (at your option)
@@ -609,6 +609,23 @@ void glReadBuffer(GLenum mode)
 	TRY();
 
 	VGLReadBuffer(mode);
+
+	CATCH();
+}
+
+
+void glReadPixels(GLint x, GLint y, GLsizei width, GLsizei height,
+	GLenum format, GLenum type, GLvoid *pixels)
+{
+	if(vglfaker::getExcludeCurrent())
+	{
+		_glReadPixels(x, y, width, height, format, type, pixels);
+		return;
+	}
+
+	TRY();
+
+	VGLReadPixels(x, y, width, height, format, type, pixels);
 
 	CATCH();
 }